Split up CheckState
