+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/investigate/CompactionChangedRecordCount.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2022-2024 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package sleeper.systemtest.suite.investigate;
+
+import sleeper.core.statestore.FileReference;
+import sleeper.core.statestore.ReplaceFileReferencesRequest;
+
+import java.util.List;
+
+public record CompactionChangedRecordCount(ReplaceFileReferencesRequest job, List<FileReference> inputFiles, FileReference outputFile) {
+    public String jobId() {
+        return job.getJobId();
+    }
+
+    public long recordsBefore() {
+        return inputFiles.stream().mapToLong(FileReference::getNumberOfRecords).sum();
+    }
+
+    public long recordsAfter() {
+        return outputFile.getNumberOfRecords();
+    }
+}
